further split TheoryForge_MFLike including now Foreground and BandPass theories
