+++ soliket/Foreground.yaml
@@ -1,5 +1,24 @@
+foregrounds:
+  normalisation:
+    nu_0: 150.0
+    ell_0: 3000
+    T_CMB: 2.725
+  components:
+    tt:
+      - kSZ
+      - tSZ_and_CIB
+      - cibp
+      - dust
+      - radio
+    te: 
+      - radio
+      - dust
+    ee: 
+      - radio
+      - dust
+
 params:
-  # Foregrounds
+ # Foregrounds
   a_tSZ:
     prior:
       min: 3.0
@@ -36,18 +55,50 @@ params:
       max: 2.4
     proposal: 0.03
     latex: \beta_c
-  n_CIBC:
-    prior:
-      min: 1.0
-      max: 1.4
-    proposal: 0.045
-    latex: n_\mathrm{CIBC}
   a_s:
     prior:
       min: 2.8
       max: 3.4
     proposal: 0.01
     latex: a_s
+  a_gtt:
+    prior:
+      dist: norm
+      loc: 2.79
+      scale: 0.45
+    proposal: 0.4
+    latex: a_\mathrm{dust}^\mathrm{TT}
+  a_gte:
+    prior:
+      dist: norm
+      loc: 0.36
+      scale: 0.04
+    proposal: 0.04
+    latex: a_\mathrm{dust}^\mathrm{TE}
+  a_gee:
+    prior:
+      dist: norm
+      loc: 0.13
+      scale: 0.03
+    proposal: 0.03
+    latex: a_\mathrm{dust}^\mathrm{EE}
+  a_psee:
+    prior:
+      min: 0
+    proposal: 0.05
+    latex: a_\mathrm{ps}^\mathrm{EE}
+  a_pste:
+    prior:
+      min: -1
+      max: 1
+    proposal: 0.05
+    latex: a_\mathrm{ps}^\mathrm{TE}
+  xi:
+    prior:
+      min: 0
+      max: 0.2
+    proposal: 0.05
+    latex: \xi
   T_d:
     prior:
       min:  8.60
