+++ soliket/__init__.py
@@ -5,6 +5,8 @@ from .clusters import ClusterLikelihood  # noqa: F401
 from .mflike import MFLike  # noqa: F401
 from .mflike import TheoryForge_MFLike
 from .xcorr import XcorrLikelihood  # noqa: F401
+from .foreground import Foreground
+from .bandpass import BandPass
 
 try:
     import pyccl as ccl  # noqa: F401
