+++ soliket/mflike/TheoryForge_MFLike.yaml
@@ -9,133 +9,12 @@ spectra:
   lmin: 0
   lmax: 7000
 
-band_integration:
-  external_bandpass: False
-  nsteps: 1
-  bandwidth: 0
-
 systematics_template:
   has_file: False
   rootname: "test_template"
 
-foregrounds:
-  normalisation:
-    nu_0: 150.0
-    ell_0: 3000
-    T_CMB: 2.725
-
-  components:
-    tt:
-      - kSZ
-      - tSZ_and_CIB
-      - cibp
-      - dust
-      - radio
-    te: 
-      - radio
-      - dust
-    ee: 
-      - radio
-      - dust
-
 params:
-  # Foregrounds
-  a_tSZ:
-    prior:
-      min: 3.0
-      max: 3.6
-    proposal: 0.05
-    latex: a_\mathrm{tSZ}
-  a_kSZ:
-    prior:
-      min: 1.4
-      max: 1.8
-    proposal: 0.1
-    latex: a_\mathrm{kSZ}
-  a_p:
-    prior:
-      min: 6.2
-      max: 7.6
-    proposal: 0.075
-    latex: a_p
-  beta_p:
-    prior:
-      min: 1.8
-      max: 2.2
-    proposal: 0.015
-    latex: \beta_p
-  a_c:
-    prior:
-      min: 4.4
-      max: 5.4
-    proposal: 0.12
-    latex: a_c
-  beta_c:
-    prior:
-      min: 2.0
-      max: 2.4
-    proposal: 0.03
-    latex: \beta_c
-  a_s:
-    prior:
-      min: 2.8
-      max: 3.4
-    proposal: 0.01
-    latex: a_s
-  a_gtt:
-    prior:
-      dist: norm
-      loc: 2.79
-      scale: 0.45
-    proposal: 0.4
-    latex: a_\mathrm{dust}^\mathrm{TT}
-  a_gte:
-    prior:
-      dist: norm
-      loc: 0.36
-      scale: 0.04
-    proposal: 0.04
-    latex: a_\mathrm{dust}^\mathrm{TE}
-  a_gee:
-    prior:
-      dist: norm
-      loc: 0.13
-      scale: 0.03
-    proposal: 0.03
-    latex: a_\mathrm{dust}^\mathrm{EE}
-  a_psee:
-    prior:
-      min: 0
-    proposal: 0.05
-    latex: a_\mathrm{ps}^\mathrm{EE}
-  a_pste:
-    prior:
-      min: -1
-      max: 1
-    proposal: 0.05
-    latex: a_\mathrm{ps}^\mathrm{TE}
-  xi:
-    prior:
-      min: 0
-      max: 0.2
-    proposal: 0.05
-    latex: \xi
-  T_d:
-    prior:
-      min:  8.60
-      max: 10.60
-    proposal: 0.6
-    latex: T_d
   # Systematics
-  bandint_shift_93:
-    value: 0
-    latex: \Delta_{\rm band}^{93}
-  bandint_shift_145:
-    value: 0
-    latex: \Delta_{\rm band}^{145}
-  bandint_shift_225:
-    value: 0
-    latex: \Delta_{\rm band}^{225}
   calT_93:
     value: 1
     latex: \mathrm{Cal}_{\rm T}^{93}
