+++ soliket/tests/test_mflike.yaml
@@ -227,6 +227,10 @@ theory:
       lens_potential_accuracy: 1
   soliket.TheoryForge_MFLike: 
     stop_at_error: True
+  soliket.Foreground:
+    stop_at_error: True
+  soliket.BandPass:
+    stop_at_error: True
 
 sampler:
   evaluate:
